Replaced size_t by unsigned int as it didn't make sense to allocate more then the size that can be stored by this data type

+++ Vec.c
@@ -11,8 +11,8 @@ static unsigned int round_next_p_two(unsigned int n) {
   return n;
 }
 
-static int alloc_vector_memory(char **data, size_t *capacity,
-                               unsigned int data_size, size_t new_capacity) {
+static int alloc_vector_memory(char **data, unsigned int *capacity,
+                               unsigned int data_size, unsigned int new_capacity) {
   void *ptr;
 
   if ((ptr = realloc(*data, new_capacity * data_size)) == NULL)
@@ -22,27 +22,27 @@ static int alloc_vector_memory(char **data, size_t *capacity,
   return 0;
 }
 
-int vector_expand(char **data, size_t *size, size_t *capacity,
+int vector_expand(char **data, unsigned int *size, unsigned int *capacity,
                   unsigned int data_size, unsigned int expanded_size) {
   if (*size + expanded_size > *capacity) {
-    size_t new_capacity = round_next_p_two(*capacity + expanded_size);
+    unsigned int new_capacity = round_next_p_two(*capacity + expanded_size);
 
     return alloc_vector_memory(data, capacity, data_size, new_capacity);
   }
   return 0;
 }
 
-int vector_reserve(char **data, size_t *capacity, unsigned int data_size,
-                   size_t new_capacity) {
+int vector_reserve(char **data, unsigned int *capacity, unsigned int data_size,
+                   unsigned int new_capacity) {
   if (new_capacity > *capacity) {
     return alloc_vector_memory(data, capacity, data_size, new_capacity);
   }
   return 0;
 }
 
-void vector_erase(char **data, size_t *size, size_t *capacity,
-                  unsigned int data_size, size_t beg, size_t end) {
-  size_t block_size = end - beg;
+void vector_erase(char **data, unsigned int *size, unsigned int *capacity,
+                  unsigned int data_size, unsigned int beg, unsigned int end) {
+  unsigned int block_size = end - beg;
 
   if (block_size >= *size) {
     free(*data);
@@ -58,7 +58,7 @@ void vector_erase(char **data, size_t *size, size_t *capacity,
   *size -= block_size;
 }
 
-void vector_insert_make_room(char **data, size_t size, unsigned int data_size, unsigned int position, unsigned int length) {
+void vector_insert_make_room(char **data, unsigned int size, unsigned int data_size, unsigned int position, unsigned int length) {
  if (position < size)
  {
      memmove(*data + (position + length) * data_size,
@@ -67,7 +67,7 @@ void vector_insert_make_room(char **data, size_t size, unsigned int data_size, u
  } 
 }
 
-void vector_insert_array(char **data, size_t *size, unsigned int data_size,
+void vector_insert_array(char **data, unsigned int *size, unsigned int data_size,
                          char *array, unsigned int length, unsigned int position) {
  vector_insert_make_room(data, *size, data_size, position, length);
  memcpy(*data + position * data_size, array, length * data_size);
